Move `brioche_packer::autowrap` into `brioche_autowrap` proper
There were a few tweaks to work better with `brioche-ld` as well
