+++ crates/brioche-ld/src/main.rs
@@ -1,6 +1,7 @@
-use std::{path::PathBuf, process::ExitCode};
+use std::{collections::HashSet, path::PathBuf, process::ExitCode};
 
 use bstr::ByteSlice as _;
+use eyre::{Context as _, OptionExt as _};
 
 enum Mode {
     AutowrapEnabled {
@@ -22,17 +23,20 @@ fn main() -> ExitCode {
     }
 }
 
-fn run() -> Result<ExitCode, LdError> {
-    let current_exe = std::env::current_exe().map_err(LdError::FailedToGetCurrentExe)?;
+fn run() -> eyre::Result<ExitCode> {
+    let current_exe = std::env::current_exe().context("failed to get current executable")?;
     let current_exe_dir = current_exe
         .parent()
-        .ok_or_else(|| LdError::FailedToGetCurrentExeDir)?;
+        .ok_or_eyre("failed to get current executable dir")?;
     let current_exe_parent_dir = current_exe_dir
         .parent()
-        .ok_or_else(|| LdError::FailedToGetCurrentExeDir)?;
+        .ok_or_eyre("failed to get current executable dir")?;
     let ld_resource_dir = current_exe_parent_dir.join("libexec").join("brioche-ld");
     if !ld_resource_dir.is_dir() {
-        return Err(LdError::LinkerResourceDirNotFound(ld_resource_dir));
+        eyre::bail!(
+            "failed to find linker resource dir: {}",
+            ld_resource_dir.display()
+        );
     }
 
     let linker = ld_resource_dir.join("ld");
@@ -44,25 +48,27 @@ fn run() -> Result<ExitCode, LdError> {
 
     let mut args = std::env::args_os().skip(1);
     while let Some(arg) = args.next() {
-        let arg = <[u8]>::from_os_str(&arg).ok_or_else(|| LdError::InvalidArg)?;
+        let arg = <[u8]>::from_os_str(&arg).ok_or_eyre("invalid arg")?;
         let arg = bstr::BStr::new(arg);
 
         if &**arg == b"-o" {
-            let output = args.next().ok_or_else(|| LdError::InvalidArg)?;
+            let output = args.next().ok_or_eyre("invalid arg")?;
             output_path = PathBuf::from(output);
         } else if let Some(output) = arg.strip_prefix(b"-o") {
-            let output = output.to_path().map_err(|_| LdError::InvalidPath)?;
+            let output = output.to_path().map_err(|_| eyre::eyre!("invalid path"))?;
             output.clone_into(&mut output_path);
         } else if &**arg == b"-L" {
-            let lib_path = args.next().ok_or_else(|| LdError::InvalidArg)?;
+            let lib_path = args.next().ok_or_eyre("invalid arg")?;
             library_search_paths.push(PathBuf::from(lib_path));
         } else if let Some(lib_path) = arg.strip_prefix(b"-L") {
-            let lib_path = lib_path.to_path().map_err(|_| LdError::InvalidPath)?;
+            let lib_path = lib_path
+                .to_path()
+                .map_err(|_| eyre::eyre!("invalid path"))?;
             library_search_paths.push(lib_path.to_owned());
         } else if arg.starts_with(b"-") {
             // Ignore other arguments
         } else {
-            let input_path = arg.to_path().map_err(|_| LdError::InvalidPath)?;
+            let input_path = arg.to_path().map_err(|_| eyre::eyre!("invalid path"))?;
             input_paths.push(input_path.to_owned());
         }
     }
@@ -74,9 +80,9 @@ fn run() -> Result<ExitCode, LdError> {
         Ok("false") => Mode::AutowrapDisabled,
         _ => {
             let resource_dir = brioche_resources::find_output_resource_dir(&output_path)
-                .map_err(LdError::ResourceDirError)?;
+                .context("error while finding resource dir")?;
             let all_resource_dirs = brioche_resources::find_resource_dirs(&current_exe, true)
-                .map_err(LdError::ResourceDirError)?;
+                .context("error while finding resource dir")?;
             Mode::AutowrapEnabled {
                 resource_dir,
                 all_resource_dirs,
@@ -106,17 +112,27 @@ fn run() -> Result<ExitCode, LdError> {
             resource_dir,
             all_resource_dirs,
         } => {
-            brioche_autowrap::autowrap(brioche_autowrap::AutowrapOptions {
-                program_path: &output_path,
-                packed_exec_path: &packed_path,
-                resource_dir: &resource_dir,
-                all_resource_dirs: &all_resource_dirs,
-                sysroot: &ld_resource_dir,
-                library_search_paths: &library_search_paths,
-                input_paths: &input_paths,
-                skip_libs: &[],
-                skip_unknown_libs,
-                runtime_library_dirs: &[],
+            let dynamic_linking_config = brioche_autowrap::DynamicLinkingConfig {
+                library_paths: library_search_paths,
+                skip_libraries: HashSet::new(),
+                extra_libraries: vec![],
+                skip_unknown_libraries: skip_unknown_libs,
+            };
+            brioche_autowrap::autowrap(&brioche_autowrap::AutowrapConfig {
+                resource_dir,
+                all_resource_dirs,
+                inputs: brioche_autowrap::AutowrapInputs::Paths(vec![output_path]),
+                quiet: true,
+                link_dependencies: vec![ld_resource_dir],
+                dynamic_binary: Some(brioche_autowrap::DynamicBinaryConfig {
+                    packed_executable: packed_path,
+                    dynamic_linking: dynamic_linking_config.clone(),
+                }),
+                shared_library: Some(brioche_autowrap::SharedLibraryConfig {
+                    dynamic_linking: dynamic_linking_config,
+                }),
+                rewrap: None,
+                script: None,
             })?;
         }
         Mode::AutowrapDisabled => {
@@ -125,30 +141,4 @@ fn run() -> Result<ExitCode, LdError> {
     };
 
     Ok(ExitCode::SUCCESS)
-}
-
-#[derive(Debug, thiserror::Error)]
-enum LdError {
-    #[error("error wrapping binary: {0}")]
-    AutowrapError(#[from] brioche_autowrap::AutowrapError),
-    #[error("invalid arg")]
-    InvalidArg,
-    #[error("invalid path")]
-    InvalidPath,
-    #[error("failed to find current executable: {0}")]
-    FailedToGetCurrentExe(#[source] std::io::Error),
-    #[error("failed to get directory containing current executable")]
-    FailedToGetCurrentExeDir,
-    #[error("brioche-ld resource dir not found (expected to be at {0})")]
-    LinkerResourceDirNotFound(PathBuf),
-    #[error("{0}")]
-    IoError(#[from] std::io::Error),
-    #[error("{0}")]
-    GoblinError(#[from] goblin::error::Error),
-    #[error("error when finding resource dir")]
-    ResourceDirError(#[from] brioche_resources::PackResourceDirError),
-    #[error("error writing packed program")]
-    InjectPackError(#[from] brioche_pack::InjectPackError),
-    #[error("error adding blob: {0}")]
-    AddBlobError(#[from] brioche_resources::AddBlobError),
 }
